Default to exiting with success if the check run is successful
Towards clarification of warden's place in the pipeline in #165 -
warden shouldn't be making the "go/no-go" call on the dataset itself
but should be generating metadata such that something else can make
the decision. As part of this, we don't want jobs to fail/abort with
error if the checks in fact ran successfully.